requirements.txt
cython
ipython
matplotlib
pandas
scikit-learn 
seaborn
pyarrow
tqdm
lightning
mkl 
rich
torchmetric

enformer_pytorch
kipoiseq
pysam
wandb
vcfpy
